requirements.txt
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0